Print ttl, creation-time, and absolute-expiry-time the dTests way
Previously, default ttl (when nothing was set) was printed as
uint.MaxValue. This commit changes it to 0 in this situation.

Previously, the values of DateTime type (or AMQP's timestamp)
were formated weirdly (Ticks == 0 was printed as negative nubmer).
This commit prints the Ticks value, which is what other clis do.

ab_diff = [{'ttl': [4294967295, 0]}]
'creation-time': -62135596800, 'ttl': 4294967295, 'absolute-expiry-time': -62135596800